requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
websockets==12.0
python-multipart==0.0.6

# For local model support (Ollama) and cloud APIs
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Production server
gunicorn==21.2.0 